move metafunctions to separate header
